08.05.2021

+++ interrupts.cpp
@@ -30,7 +30,7 @@ picSlaveData(0xA1)
     uint16_t CodeSegment = gdt->CodeSegmentDescriptor();
     const uint8_t IDT_INTERRUPT_GATE = 0xE;
 
-    for (int i = 0; i < 256; i++) {
+    for (uint16_t i = 0; i < 256; i++) {
         SetInterruptDescriptorTableEntry(i, CodeSegment, &IgnoreInterruptRequest, 0, IDT_INTERRUPT_GATE);
     }
     SetInterruptDescriptorTableEntry(0x20, CodeSegment, &HandleInterruptRequest0x00, 0, IDT_INTERRUPT_GATE);
